Moved file index.js to Carousel.js
Added a new boolean property to show the indicators at the bottom or top
	indicatorAtBottom={false} // Show at the top

Added new styles for the  indicator positions
	pageIndicatorTop
	pageIndicatorBottom

Clicking the indicators will switch the card/page it is on.

Set the background of the indicator to transparent.

Set bounces=false on the ScrollView used for the Carousel